% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/OnDiskVector-class.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{length,OnDiskVector-method}
\alias{[,OnDiskVector,numeric,missing-method}
\alias{[,OnDiskVector,missing,missing-method}
\alias{[<-,OnDiskVector,missing,missing-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{length}{OnDiskVector}(x)

\S4method{[}{OnDiskVector,numeric,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{OnDiskVector,missing,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{OnDiskVector,missing,missing}(x, i, j, ...) <- value
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
