\name{MALDIquant-internal}

\alias{[,MassPeaks,logical,missing-method}
\alias{[,MassPeaks,numeric,missing-method}

\title{Internal MALDIquant Functions}
\description{
Internal MALDIquant functions.
}
\note{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written).
}
\keyword{internal}

