% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options-functions.R
\name{MALDIquantOptions}
\alias{MALDIquantOptions}
\alias{.negativeIntensitiesAllowed}
\title{MALDIquant options}
\usage{
MALDIquantOptions()

.negativeIntensitiesAllowed()
}
\value{
A \code{list} of MALDIquant options and the single option
values for the individual accessors.
}
\description{
MALDIquant defines a few options globally using the standard R
options mechanism. The current values of these options can be
queried with \code{MALDIquantOptions()}. The options are:
}
\details{
\itemize{
\item{\code{allowNegativeIntensities}: flag, that controls whether negative
intensities are allowed. Please note that negative intensities could
badly influence the results of some algorithms e.g. baseline correction,
smoothing, \dots. The default value is \code{FALSE}}
}
}
\examples{
# save old options
opt <- options()

# set MALDIquant options
options(MALDIquant=list(allowNegativeIntensities=FALSE))

MALDIquantOptions()

# reset old options
options(opt)
}
