\name{plotMsiSlice-methods}
\docType{methods}
\title{Plots a Mass Spectrometry Imaging dataset.}
\alias{plotMsiSlice}
\alias{plotMsiSlice,list-method}
\alias{plotMsiSlice,array-method}
\alias{plotMsiSlice,matrix-method}
\description{
This function allows to plot a slice of a mass spectrometry imaging dataset
represented by a \code{list} of \code{\linkS4class{AbstractMassObject}} objects
or an \code{array} or a \code{matrix}.
}
\usage{
\S4method{plotMsiSlice}{list}(x, center, tolerance,
  colRamp=colorRamp(c("black", "blue", "green", "yellow", "red")),
  interpolate=FALSE, scale=TRUE, legend=TRUE, \dots)
\S4method{plotMsiSlice}{array}(x,
  colRamp=colorRamp(c("black", "blue", "green", "yellow", "red")),
  interpolate=FALSE, scale=TRUE, legend=TRUE, sub=NULL, \dots)
\S4method{plotMsiSlice}{matrix}(x,
  colRamp=colorRamp(c("black", "blue", "green", "yellow", "red")),
  interpolate=FALSE, scale=TRUE, legend=TRUE, \dots)
}
\arguments{
  \item{x}{The mass spectrometry imaging dataset. It could be a \code{list} of
    \code{\linkS4class{MassSpectrum}}/\code{\linkS4class{MassPeaks}} objects or
    an \code{array} (e.g. generated by \code{\link[MALDIquant]{msiSlices}}) or
    a \code{matrix}.}
  \item{center}{\code{double}, if \code{x} is a \code{list} of
    \code{\linkS4class{MassSpectrum}}/\code{\linkS4class{MassPeaks}} objects
    this argument represent the \emph{center} mass value of the slices,
    see \code{\link[MALDIquant]{msiSlices}} for details.}
  \item{tolerance}{\code{double}, if \code{center} is given \code{tolerance}
    specifies the thickness of the slices (\code{center + c(-tolerance,
    tolerance)}), see \code{\link[MALDIquant]{msiSlices}} for details.}
  \item{colRamp}{colours as \code{colorRamp} \code{function}, see
    \code{\link[grDevices]{colorRamp}} for details.}
  \item{interpolate}{\code{logical}, use linear interpolation when drawing the
    image, see \code{\link[graphics]{rasterImage}} for details.}
  \item{scale}{\code{logical}, if \code{TRUE} all values are divided by the
    maximal value of the slice to get values between 0 and 1.}
  \item{legend}{\code{logical}, if \code{TRUE} a reference color gradient is
    plotted on the right hand side of the plot. The upper color represents the
    highest value in the slice and the lower color the lowest value
    respectively.}
  \item{sub}{\code{character}, sub title for the plot, see
    \code{\link[graphics]{title}}.}
  \item{\dots}{arguments to be passed to \code{\link[graphics]{plot}}.}
}
\details{
Each \code{\linkS4class{MassSpectrum}}/\code{\linkS4class{MassPeaks}} object in
\code{x} must contain a \code{list} named \code{imaging} with an element
\code{pos} that stores the \code{x} and \code{y} value of the spectrum, e.g.:
\preformatted{
> metaData(spectra[[1]])$imaging$pos
x y
1 5
}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\seealso{
\code{\linkS4class{AbstractMassObject}},
\code{\linkS4class{MassSpectrum}},
\code{\linkS4class{MassPeaks}},
\code{\link[MALDIquant]{msiSlices}},
\code{\link[MALDIquant]{plot,MassSpectrum,missing-method}}

Website: \url{http://strimmerlab.org/software/maldiquant/}

Example:
\url{https://github.com/sgibb/MALDIquantExamples/blob/master/vignettes/ims.Rnw}

Shiny example: \url{https://github.com/sgibb/ims-shiny/}
}
\examples{
\dontrun{
plotMsiSlice(spectra, center=3362.3, tolerance=0.25)
}
}
\keyword{Graphics|hplot}
