\name{sA1}
\docType{data}
\alias{sA1}
\title{Example Mass Spectra}
\description{
This dataset contains two example mass spectra (they are technical replicates).
It is used to demonstrate the usage of \code{\link{MALDIquant-package}}.
}
\usage{sA1}
\format{A list containing two \code{\link[MALDIquant]{MassSpectrum-class}}
    objects.}
\seealso{
    \code{\link[MALDIquant]{MassSpectrum-class}}
}
\keyword{datasets}
