\name{MALDIquant-deprecated}
\alias{MALDIquant-deprecated}
%------ PLEASE: put \alias{.} here for EACH !
%------ NOTE: ../R/Deprecated.R must be synchronized with this!
\title{Deprecated Functions/Methods in Package \pkg{MALDIquant}}
\description{
  These functions/methods are provided for compatibility with older versions
  \code{\link{MALDIquant}} only, and may be defunct as soon as the next release.
}
%\details{
%\describe{
%  % since MALDIquant x.y.z
%  %\item{function}{explanation}
%  % since MALDIquant 1.11.12
%  }
%}
\seealso{
  \code{\link{Deprecated}}
}
\keyword{internal}
\keyword{misc}
