\name{MALDIquant-internal}

\alias{approxfun}
\alias{show}

\alias{.binPeaks}
\alias{.calculateLabelPositions}
\alias{.doByLabels}
\alias{.filterPeaks}
\alias{.findLocalMaxima}
\alias{.grouperRelaxedHighestAtReference}
\alias{.grouperStrict}
\alias{.isArgument}
\alias{.isFunctionList}
\alias{.labelOverlap}
\alias{.lowerConvexHullR}
\alias{.merge}
\alias{.mergeMassPeaks}
\alias{.mergeMassSpectra}
\alias{.removeArguments}
\alias{.replaceNonEmptyMassObjects}
\alias{.snipR}
\alias{.stopIfNotMassObjectList}
\alias{.stopIfNotMassPeaksList}
\alias{.stopIfNotMassSpectrumList}
\alias{.testLabelOverlap}
\alias{.textLabelRects}
\alias{.validAbstractMassObject}
\alias{.warp}
\alias{.warpingFunctionCubic}
\alias{.warpingFunctionLowess}
\alias{.which.nearby}

\title{Internal MALDIquant Functions}
\description{
Internal MALDIquant functions.
}
\note{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written).
}
\keyword{internal}

