\name{MALDIquant-internal}

\alias{approxfun}
\alias{show}
\alias{[,MassPeaks,logical,missing-method}
\alias{[,MassPeaks,numeric,missing-method}

\alias{.averageMassSpectra}
\alias{.binPeaks}
\alias{.calibrateIntensitySimple}
\alias{.calibrateProbabilisticQuotientNormalization}
\alias{.calculateLabelPositions}
\alias{.colMedians}
\alias{.dilation}
\alias{.deprecated}
\alias{.doByLabels}
\alias{.erosion}
\alias{.filterPeaks}
\alias{.filter}
\alias{.findLocalMaxima}
\alias{.findLocalMaximaLogical}
\alias{.grouperRelaxedHighestAtReference}
\alias{.grouperStrict}
\alias{.irregularScore}
\alias{.isFunctionList}
\alias{.isMassObject}
\alias{.isMassObjectList}
\alias{.labelOverlap}
\alias{.localMaxima}
\alias{.lowerConvexHullR}
\alias{.match.arg}
\alias{.memoryUsageStr}
\alias{.mergeMassPeaks}
\alias{.mergeMetaData}
\alias{.prepareShow}
\alias{.prepareShowGroupName}
\alias{.savitzkyGolayCoefficients}
\alias{.snipR}
\alias{.stopIfNotIsMassObjectList}
\alias{.stopIfNotIsMassPeaksList}
\alias{.stopIfNotIsMassSpectrumList}
\alias{.stopIfNotIsValidHalfWindowSize}
\alias{.testLabelOverlap}
\alias{.textLabelRects}
\alias{.topHatR}
\alias{.transformIntensity}
\alias{.validAbstractMassObject}
\alias{.validMassPeaks}
\alias{.warp}
\alias{.warpingFunctionCubic}
\alias{.warpingFunctionLowess}
\alias{.which.closest}

\title{Internal MALDIquant Functions}
\description{
Internal MALDIquant functions.
}
\note{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written).
}
\keyword{internal}

