\name{MALDIquant-internal}

\alias{approxfun}
\alias{show}
\alias{[,MassPeaks,logical,missing-method}
\alias{[,MassPeaks,numeric,missing-method}

\alias{.binPeaks}
\alias{.calculateLabelPositions}
\alias{.dilation}
\alias{.doByLabels}
\alias{.erosion}
\alias{.filterPeaks}
\alias{.filter}
\alias{.findLocalMaxima}
\alias{.findLocalMaximaLogical}
\alias{.grouperRelaxedHighestAtReference}
\alias{.grouperStrict}
\alias{.iplot}
\alias{.iplotSingle}
\alias{.iplotUsage}
\alias{.irregularScore}
\alias{.isFunctionList}
\alias{.lim}
\alias{.labelOverlap}
\alias{.localMaxima}
\alias{.lowerConvexHullR}
\alias{.memoryUsageStr}
\alias{.merge}
\alias{.mergeMassPeaks}
\alias{.mergeMassSpectra}
\alias{.mergeMetaData}
\alias{.moveLim}
\alias{.moveXlim}
\alias{.openInteractiveDevice}
\alias{.prepareShow}
\alias{.prepareShowGroupName}
\alias{.replaceNonEmptyMassObjects}
\alias{.savitzkyGolayCoefficients}
\alias{.snipR}
\alias{.stopIfNotIsMassObjectList}
\alias{.stopIfNotIsMassPeaksList}
\alias{.stopIfNotIsMassSpectrumList}
\alias{.stopIfNotIsValidHalfWindowSize}
\alias{.testLabelOverlap}
\alias{.textLabelRects}
\alias{.topHatR}
\alias{.validAbstractMassObject}
\alias{.validMassPeaks}
\alias{.warp}
\alias{.warpingFunctionCubic}
\alias{.warpingFunctionLowess}
\alias{.which.closest}
\alias{.xlim}
\alias{.ylim}
\alias{.zoomLim}
\alias{.zoomXlim}
\alias{.zoomYlim}

\title{Internal MALDIquant Functions}
\description{
Internal MALDIquant functions.
}
\note{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written).
}
\keyword{internal}

