\name{MALDIquant-package}
\alias{MALDIquant-package}
\alias{MALDIquant}
\docType{package}
\title{
Quantitative analysis of MALDI-TOF MS data
}
\description{
This package should provide an easy framework for quantitative analysis of 
matrix-assisted laser desorption/ionization - time-of-flight (MALDI-TOF) 
mass spectrometry data.
Up to now only importing and plotting of a single mass spectrum are supported.
}
\details{
\tabular{ll}{
Package: \tab MALDIquant\cr
Version: \tab 0.1\cr
Date: \tab 2011-02-22\cr
License: \tab GPL (>= 3)\cr
URL: \tab http://www.sebastiangibb.de/maldiquant\cr
}

Main classes:

\code{\linkS4class{SingleSpectrum}}: Represents a single spectrum.
}

\author{
Sebastian Gibb

Maintainer: Sebastian Gibb <mail@sebastiangibb.de>
}
\references{
See website: \url{http://www.sebastiangibb.de/maldiquant}
}
\keyword{package}
\seealso{
\code{\linkS4class{SingleSpectrum}}
}
