\name{pA1}
\docType{data}
\alias{pA1}
\title{Example Mass Spectra (extracted peaks)}
\description{
This dataset contains two example peak lists (they are technical
replicates). This peak lists are extracted out of \code{\link[MALDIquant]{sA1}}.
It is used to demonstrate the usage of \code{\link{MALDIquant-package}}.
}
\usage{pA1}
\format{A list containing two \code{\link[MALDIquant]{MassPeaks-class}}
    objects.}
\seealso{
    \code{\link[MALDIquant]{MassPeaks-class}},
    \code{\link[MALDIquant]{sA1}}
}
\keyword{datasets}
